almost implemented the rotatable camera, but its still quite buggy, trying to stamp the weirdness out

+++ Deltaglider.cpp
@@ -373,7 +373,7 @@ double DeltaGlider::Get_total_mass()
 	return net_mass;
 }
 
-double DeltaGlider::Get_PMI()
+long double DeltaGlider::Get_PMI()
 {	return PMI;
 }
